refact(dir): Move state files to src/store
