+++ src/routes/Login/index.tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from 'react'
 import View from './view'
 import { useDispatch, useSelector } from 'react-redux'
 import { RootState } from '../../store'
-import { setPending } from '../../state'
+import { setPending } from '../../store/status'
 
 const selector = ({ result }: RootState) => ({ result })
